editor/libeditor/tests/test_nsIPlaintextEditor_insertLineBreak.html

<!DOCTYPE>
<html>
<head>
  <title>Test for nsIPlaintextEditor.insertLineBreak()</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<div id="display">
</div>
<input value="abcdef">
<textarea>abcdef</textarea>
<div id="content" contenteditable></div>
<pre id="test">
</pre>

<script class="testbody" type="application/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  let input = document.getElementsByTagName("input")[0];
  let textarea = document.getElementsByTagName("textarea")[0];
  let contenteditable = document.getElementById("content");
  let selection = window.getSelection();

  let inputEvents = [];
  function onInput(event) {
    inputEvents.push(event);
  }

  input.focus();
  input.selectionStart = input.selectionEnd = 3;
  inputEvents = [];
  input.addEventListener("input", onInput);
  try {
    getPlaintextEditor(input).insertLineBreak();
  } catch (e) {}
  input.removeEventListener("input", onInput);
  is(input.value, "abcdef", "nsIPlaintextEditor.insertLineBreak() should do nothing on single line editor");
  is(inputEvents.length, 0, "nsIPlaintextEditor.insertLineBreak() shouldn't cause 'input' event on single line editor");

  textarea.focus();
  textarea.selectionStart = textarea.selectionEnd = 3;
  inputEvents = [];
  textarea.addEventListener("input", onInput);
  getPlaintextEditor(textarea).insertLineBreak();
  textarea.removeEventListener("input", onInput);
  is(textarea.value, "abc\ndef", "nsIPlaintextEditor.insertLineBreak() should insert \n into multi-line editor");
  is(inputEvents.length, 1, "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on multi-line editor");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (on multi-line editor)');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (on multi-line editor)');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (on multi-line editor)');
  is(inputEvents[0].inputType, "insertLineBreak",
     'inputType should be "insertLineBreak" on multi-line editor');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertLineBreak" on multi-line editor');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertLineBreak" on multi-line editor');

  // Note that despite of the name, insertLineBreak() should insert paragraph separator in HTMLEditor.

  document.execCommand("defaultParagraphSeparator", false, "br");

  contenteditable.innerHTML = "abcdef";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "abc<br>def",
     "nsIPlaintextEditor.insertLineBreak() should insert <br> element into text node when defaultParagraphSeparator is \"br\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has only text node when defaultParagraphSeparator is \"br\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "br") #1');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "br") #1');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "br") #1');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #1');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #1');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #1');

  contenteditable.innerHTML = "<p>abcdef</p>";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "<p>abc</p><p>def</p>",
     "nsIPlaintextEditor.insertLineBreak() should add <p> element after <p> element even when defaultParagraphSeparator is \"br\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has <p> element when defaultParagraphSeparator is \"br\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "br") #2');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "br") #2');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "br") #2');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #2');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #2');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #2');

  contenteditable.innerHTML = "<div>abcdef</div>";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "<div>abc<br>def</div>",
     "nsIPlaintextEditor.insertLineBreak() should insert <br> element into <div> element when defaultParagraphSeparator is \"br\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has <div> element when defaultParagraphSeparator is \"br\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "br") #3');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "br") #3');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "br") #3');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #3');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #3');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #3');

  contenteditable.innerHTML = "<pre>abcdef</pre>";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "<pre>abc<br>def</pre>",
     "nsIPlaintextEditor.insertLineBreak() should insert <br> element into <pre> element when defaultParagraphSeparator is \"br\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has <pre> element when defaultParagraphSeparator is \"br\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "br") #4');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "br") #4');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "br") #4');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #4');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #4');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "br") #4');

  document.execCommand("defaultParagraphSeparator", false, "p");

  contenteditable.innerHTML = "abcdef";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "<p>abc</p><p>def</p>",
     "nsIPlaintextEditor.insertLineBreak() should create <p> elements when there is only text node and defaultParagraphSeparator is \"p\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has only text node when defaultParagraphSeparator is \"p\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "p") #1');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "p") #1');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "p") #1');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #1');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #1');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #1');

  contenteditable.innerHTML = "<p>abcdef</p>";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "<p>abc</p><p>def</p>",
     "nsIPlaintextEditor.insertLineBreak() should add <p> element after <p> element when defaultParagraphSeparator is \"p\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has <p> element when defaultParagraphSeparator is \"p\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "p") #2');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "p") #2');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "p") #2');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #2');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #2');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #2');

  contenteditable.innerHTML = "<div>abcdef</div>";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "<div>abc</div><div>def</div>",
     "nsIPlaintextEditor.insertLineBreak() should add <div> element after <div> element even when defaultParagraphSeparator is \"p\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has <div> element when defaultParagraphSeparator is \"p\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "p") #3');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "p") #3');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "p") #3');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #3');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #3');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #3');

  contenteditable.innerHTML = "<pre>abcdef</pre>";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "<pre>abc<br>def</pre>",
     "nsIPlaintextEditor.insertLineBreak() should insert <br> element into <pre> element when defaultParagraphSeparator is \"p\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has <pre> element when defaultParagraphSeparator is \"p\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "p") #4');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "p") #4');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "p") #4');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #4');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #4');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "p") #4');

  document.execCommand("defaultParagraphSeparator", false, "div");

  contenteditable.innerHTML = "abcdef";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "<div>abc</div><div>def</div>",
     "nsIPlaintextEditor.insertLineBreak() should create <div> elements when there is only text node and defaultParagraphSeparator is \"div\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has only text node when defaultParagraphSeparator is \"div\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "div") #1');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "div") #1');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "div") #1');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #1');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #1');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #1');

  contenteditable.innerHTML = "<p>abcdef</p>";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "<p>abc</p><p>def</p>",
     "nsIPlaintextEditor.insertLineBreak() should add <p> element after <p> element even when defaultParagraphSeparator is \"div\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has <p> element when defaultParagraphSeparator is \"div\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "div") #2');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "div") #2');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "div") #2');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #2');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #2');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #2');

  contenteditable.innerHTML = "<div>abcdef</div>";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "<div>abc</div><div>def</div>",
     "nsIPlaintextEditor.insertLineBreak() should add <div> element after <div> element when defaultParagraphSeparator is \"div\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has <div> element when defaultParagraphSeparator is \"div\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "div") #3');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "div") #3');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "div") #3');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #3');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #3');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #3');

  contenteditable.innerHTML = "<pre>abcdef</pre>";
  contenteditable.focus();
  contenteditable.scrollTop;
  selection.collapse(contenteditable.firstChild.firstChild, 3);
  inputEvents = [];
  contenteditable.addEventListener("input", onInput);
  getPlaintextEditor(contenteditable).insertLineBreak();
  contenteditable.removeEventListener("input", onInput);
  is(contenteditable.innerHTML, "<pre>abc<br>def</pre>",
     "nsIPlaintextEditor.insertLineBreak() should insert <br> element into <pre> element when defaultParagraphSeparator is \"div\"");
  is(inputEvents.length, 1,
     "nsIPlaintextEditor.insertLineBreak() should cause 'input' event once on contenteditable which has <pre> element when defaultParagraphSeparator is \"div\"");
  ok(inputEvents[0] instanceof InputEvent,
     '"input" event should be dispatched with InputEvent interface (when defaultParagraphSeparator is "div") #4');
  is(inputEvents[0].cancelable, false,
     '"input" event should be never cancelable even if "click" event (when defaultParagraphSeparator is "div") #4');
  is(inputEvents[0].bubbles, true,
     '"input" event should always bubble (when defaultParagraphSeparator is "div") #4');
  is(inputEvents[0].inputType, "insertParagraph",
     'inputType should be "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #4');
  is(inputEvents[0].data, null,
     'data should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #4');
  is(inputEvents[0].dataTransfer, null,
     'dataTransfer should be null if inputType is "insertParagraph" on HTMLEditor (when defaultParagraphSeparator is "div") #4');

  SimpleTest.finish();
});

function getPlaintextEditor(aEditorElement) {
  let editor = aEditorElement ? SpecialPowers.wrap(aEditorElement).editor : null;
  if (!editor) {
    editor = SpecialPowers.wrap(window).docShell.editingSession.getEditorForWindow(window);
  }
  return editor.QueryInterface(SpecialPowers.Ci.nsIPlaintextEditor);
}

</script>
</body>

</html>
